requirements.txt
mysql-connector-python
aiohttp
requests
pytest
